added project 3 folder
added project 3 folder that has all my code and screenshots. This that allow the user to sign up, login, log an exercise that calculates the exercise by their gender, age,weight, their heart rate and time spent on the exercise. users can edit their profile and view their profile with the most recent exercises and they can delete it from the view profile as well.and the program does allow for multiple users.