this is my Capstone project
this is my capstone project for PHP Web Development with MySQL. Its an application that focuses on users and their video game collection. it has a navmenu, uses bootstrap, multiple user functionality. users can add a game that has based on title, they date they purchased the game, they status of the game, and the rating they give the game on a scale from 1-10. Once they added it it displays it via an HTML table. they can also edit and delete a game as well.